requirements.txt
annotated-types==0.7.0
anyio==4.9.0
blinker==1.9.0
branca==0.7.2
certifi==2025.4.26
charset-normalizer==3.4.2
click==8.1.8
contourpy==1.3.2
cycler==0.12.1
elevenlabs==1.57.0
filelock==3.18.0
Flask==3.1.0
flask-cors==5.0.1
folium==0.17.0
fonttools==4.57.0
fsspec==2025.3.2
h11==0.16.0
httpcore==1.0.9
httpx==0.28.1
huggingface-hub==0.29.3
idna==3.10
inquirerpy==0.3.4
itsdangerous==2.2.0
Jinja2==3.1.6
kiwisolver==1.4.8
MarkupSafe==3.0.2
matplotlib==3.10.1
mpmath==1.3.0
networkx==3.4.2
numpy==2.2.5
opencv-python==4.11.0.86
packaging==25.0
pandas==2.2.3
pfzy==0.3.4
pillow==11.2.1
prompt_toolkit==3.0.50
psutil==7.0.0
py-cpuinfo==9.0.0
pydantic==2.11.4
pydantic_core==2.33.2
pyparsing==3.2.3
python-dateutil==2.9.0.post0
python-dotenv==1.1.0
pytz==2025.2
PyYAML==6.0.2
requests==2.32.3
scipy==1.15.2
seaborn==0.13.2
setuptools==80.1.0
six==1.17.0
sniffio==1.3.1
sympy==1.14.0
torch==2.7.0
torchvision==0.22.0
tqdm==4.67.1
typing-inspection==0.4.0
typing_extensions==4.13.2
tzdata==2025.2
ultralytics==8.3.124
ultralytics-thop==2.0.14
urllib3==2.4.0
wcwidth==0.2.13
websockets==15.0.1
Werkzeug==3.1.3
xyzservices==2024.6.0
